#803 [REFACTOR] 스토리 파일들을 stories 폴더로 이동하여 구조 정리
