+++ src/stories/Icon.mdx
@@ -12,7 +12,7 @@ import {
 } from '@storybook/blocks';
 
 import * as IconStories from './Icon.stories';
-import Icon from './Icon';
+import { Icon } from '@/shared/component';
 import {
   STATIC_ICON_LIST,
   FILL_ICON_LIST,
